test(e2e): 未読消化シナリオを article-browse から分離する
article-browse.mobile.test.ts はファイル名が記事閲覧を示すため、同居していた
「ログインから未読消化までの主要フロー」を専用ファイル inbox-digestion.test.ts へ切り出す。
直列実行であることを踏まえ、デバイス種別ではなくシナリオ単位でファイルを構成し、
モバイル向けの describe を内側にネストする。InboxPage の POM はそのまま流用する。
